spark/requirements.txt
pyspark==3.5.0
delta-spark==2.4.0
minio==7.2.0
boto3==1.34.0
mlflow==2.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
